Added transmit and filter options for graphics, to be used by povray
Optional arguments of `transmit` and `filter` to Graphics objects like
Arrow, Cylinder, etc. can be provided, to be used only by povray at the
moment. It can also be used in `plotmesh`, `plotselection` or
`plotfield`. Documentation is added for all the Graphics objects. An
example file is added under povray to test the transmit and filter
options.