Added check for field shape and dimensions
VTK currently only supports export of scalar or 2D/3D vector (column matrix) fields that live on the vertices ( shape `[1,0,0]`). Here, we add a check for this and abort the export if the field is not supported. Earlier, the code was silently exporting the first 3 components if we pass a higher dimensional vector.

We also add tests and clarify this in the documentation.